refactor: team.js 모듈로 나누기
