+++ src/Team/removeTeam.js
@@ -1,5 +1,4 @@
-import { loadTeamList } from './Team.js';
-import { $teamListContainer } from '../dom.js';
+import { loadTeamList, clearAllList } from './Team.js';
 
 export const removeTeam = async (target) => {
   if (!confirm('정말 이 팀을 삭제하시겠습니까?')) {
@@ -10,11 +9,5 @@ export const removeTeam = async (target) => {
   await api.deleteTeam(teamId);
 
   clearAllList();
-  loadTeamsList();
-};
-
-const clearAllList = () => {
-  while ($teamListContainer.firstChild) {
-    $teamListContainer.lastChild.remove();
-  }
+  loadTeamList();
 };
